created individual icon story files with preview stories for icongallery
